Migrate and restructure trending
