+++ src3/reducers/root_reducer.js
@@ -5,7 +5,7 @@ import queryReducer from './query_reducer';
 import recommendedReducer from './recommended_reducer';
 import searchResultReducer from './search_result_reducer';
 import subscriptionsReducer from './subscriptions_reducer';
-import trendingReducer from './trending_reducer';
+// import trendingReducer from './trending_reducer';
 // import userReducer from './user_reducer';
 // import settingReducer from './setting_reducer';
 import playlistsReducer from './playlists_reducer';
@@ -18,7 +18,7 @@ const rootReducer = combineReducers({
   recommended: recommendedReducer,
   searchResult: searchResultReducer,
   subscriptions: subscriptionsReducer,
-  trending: trendingReducer,
+  // trending: trendingReducer,
   // user: userReducer,
   // setting: settingReducer,
   playlists: playlistsReducer,
