+++ src3/reducers/trending_reducer.js
@@ -1,23 +1,23 @@
-import { merge } from 'lodash';
-import { RECEIVE_TRENDING } from "../actions/youtube_video_actions";
-
-let _defaultState = {
-  date: null,
-  videos: null
-};
-
-const trendingReducer = (state = _defaultState, action) => {
-  Object.freeze(state);
-
-  switch(action.type) {
-    case RECEIVE_TRENDING:
-      return {
-        date: Date.now(),
-        videos: action.videos
-      };
-    default:
-      return state;
-  }
-};
-
-export default trendingReducer;
+// import { merge } from 'lodash';
+// import { RECEIVE_TRENDING } from "../actions/youtube_video_actions";
+//
+// let _defaultState = {
+//   date: null,
+//   videos: null
+// };
+//
+// const trendingReducer = (state = _defaultState, action) => {
+//   Object.freeze(state);
+//
+//   switch(action.type) {
+//     case RECEIVE_TRENDING:
+//       return {
+//         date: Date.now(),
+//         videos: action.videos
+//       };
+//     default:
+//       return state;
+//   }
+// };
+//
+// export default trendingReducer;
