+++ src3/util/youtube_video_util/trending_util.js
@@ -1,12 +1,12 @@
-import * as YoutubeApi from '../youtube_api';
-
-export const fetchTrending = () => {
-  let baseUrl = `https://www.googleapis.com/youtube/v3/videos`;
-
-  let params = {
-    part: 'snippet,statistics,contentDetails',
-    chart: 'mostPopular',
-  };
-
-  return YoutubeApi.videos(params);
-};
+// import * as YoutubeApi from '../youtube_api';
+//
+// export const fetchTrending = () => {
+//   let baseUrl = `https://www.googleapis.com/youtube/v3/videos`;
+//
+//   let params = {
+//     part: 'snippet,statistics,contentDetails',
+//     chart: 'mostPopular',
+//   };
+//
+//   return YoutubeApi.videos(params);
+// };
