--- src3/youtube.jsx
@@ -1,23 +1,0 @@
-// /* global Promise */
-//
-// import React from 'react';
-// import ReactDOM from 'react-dom';
-// import configureStore from './store/store';
-// import Root from './components/root';
-// import { refreshToken } from './util/oauth_util';
-//
-// document.addEventListener('DOMContentLoaded', () => {
-//   const root = document.getElementById('root');
-//
-//   Promise.all([refreshToken()]).then(() => {
-//     let preloadedState = {
-//       user: JSON.parse(localStorage.getItem('google-user'))
-//     };
-//
-//     const store = configureStore(preloadedState);
-//
-//     ReactDOM.render(<Root store={store}/>, root);
-//     window.store = store;
-//     window.s = store.getState;
-//   });
-// });
